feat: Refactor controllers into domain-specific packages
Controllers and related tests have been reorganized into domain-specific packages such as `auth`, `game`, and `user`. This improves project structure by grouping related components together, enhancing maintainability and clarity. Additionally, removed unused JavaMailSenderMock and associated stack configuration.
