+++ lsadf_api/src/main/java/com/lsadf/application/game/characteristics/CharacteristicsControllerImpl.java
@@ -13,19 +13,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.lsadf.application.controllers.impl;
+package com.lsadf.application.game.characteristics;
 
 import static com.lsadf.core.infra.web.config.auth.TokenUtils.getUsernameFromJwt;
 import static com.lsadf.core.infra.web.responses.ResponseUtils.generateResponse;
 
-import com.lsadf.application.controllers.CurrencyController;
-import com.lsadf.core.application.game.currency.CurrencyService;
+import com.lsadf.core.application.game.characteristics.CharacteristicsService;
 import com.lsadf.core.application.game.game_save.GameSaveService;
-import com.lsadf.core.domain.game.currency.Currency;
+import com.lsadf.core.domain.game.characteristics.Characteristics;
 import com.lsadf.core.infra.cache.services.CacheService;
 import com.lsadf.core.infra.web.controllers.BaseController;
-import com.lsadf.core.infra.web.requests.game.currency.CurrencyRequest;
-import com.lsadf.core.infra.web.requests.game.currency.CurrencyRequestModelMapper;
+import com.lsadf.core.infra.web.requests.game.characteristics.CharacteristicsRequest;
+import com.lsadf.core.infra.web.requests.game.characteristics.CharacteristicsRequestModelMapper;
 import com.lsadf.core.infra.web.responses.GenericResponse;
 import lombok.extern.slf4j.Slf4j;
 import org.slf4j.Logger;
@@ -35,51 +34,51 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.web.bind.annotation.RestController;
 
-/** Implementation of the Currency Controller */
 @RestController
 @Slf4j
-public class CurrencyControllerImpl extends BaseController implements CurrencyController {
-
+public class CharacteristicsControllerImpl extends BaseController
+    implements CharacteristicsController {
   private final GameSaveService gameSaveService;
-  private final CurrencyService currencyService;
+  private final CharacteristicsService characteristicsService;
   private final CacheService cacheService;
 
-  private final CurrencyRequestModelMapper requestModelMapper;
+  private final CharacteristicsRequestModelMapper mapper;
 
   @Autowired
-  public CurrencyControllerImpl(
+  public CharacteristicsControllerImpl(
       GameSaveService gameSaveService,
-      CurrencyService currencyService,
+      CharacteristicsService characteristicsService,
       CacheService cacheService,
-      CurrencyRequestModelMapper requestModelMapper) {
+      CharacteristicsRequestModelMapper mapper) {
     this.gameSaveService = gameSaveService;
-    this.currencyService = currencyService;
+    this.characteristicsService = characteristicsService;
     this.cacheService = cacheService;
-    this.requestModelMapper = requestModelMapper;
+    this.mapper = mapper;
   }
 
   /** {@inheritDoc} */
   @Override
-  public ResponseEntity<GenericResponse<Void>> saveCurrency(
-      Jwt jwt, String gameSaveId, CurrencyRequest currencyRequest) {
+  public ResponseEntity<GenericResponse<Void>> saveCharacteristics(
+      Jwt jwt, String gameSaveId, CharacteristicsRequest characteristicsRequest) {
     validateUser(jwt);
     String userEmail = getUsernameFromJwt(jwt);
     gameSaveService.checkGameSaveOwnership(gameSaveId, userEmail);
 
-    Currency currency = requestModelMapper.mapToModel(currencyRequest);
-    currencyService.saveCurrency(gameSaveId, currency, cacheService.isEnabled());
+    Characteristics characteristics = mapper.mapToModel(characteristicsRequest);
+    characteristicsService.saveCharacteristics(
+        gameSaveId, characteristics, cacheService.isEnabled());
 
     return generateResponse(HttpStatus.OK);
   }
 
   /** {@inheritDoc} */
   @Override
-  public ResponseEntity<GenericResponse<Void>> getCurrency(Jwt jwt, String gameSaveId) {
+  public ResponseEntity<GenericResponse<Void>> getCharacteristics(Jwt jwt, String gameSaveId) {
     validateUser(jwt);
     String userEmail = getUsernameFromJwt(jwt);
     gameSaveService.checkGameSaveOwnership(gameSaveId, userEmail);
-    Currency currency = currencyService.getCurrency(gameSaveId);
-    return generateResponse(HttpStatus.OK, currency);
+    Characteristics characteristics = characteristicsService.getCharacteristics(gameSaveId);
+    return generateResponse(HttpStatus.OK, characteristics);
   }
 
   /** {@inheritDoc} */
