+++ lsadf_api/src/main/java/com/lsadf/application/game/stage/StageController.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.lsadf.application.controllers;
+package com.lsadf.application.game.stage;
 
 import static com.lsadf.core.infra.web.controllers.ControllerConstants.STAGE;
 import static com.lsadf.core.infra.web.controllers.ControllerConstants.Swagger.Authentications.BEARER_AUTHENTICATION;
