+++ lsadf_api/src/main/java/com/lsadf/config/LsadfConfiguration.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.lsadf.application.config;
+package com.lsadf.config;
 
 import com.lsadf.core.application.ApplicationServiceConfiguration;
 import com.lsadf.core.infra.cache.config.NoValkeyCacheConfiguration;
