+++ lsadf_api/src/main/java/com/lsadf/config/LsadfSecurityConfiguration.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.lsadf.application.config;
+package com.lsadf.config;
 
 import static com.lsadf.core.infra.web.config.security.SecurityConfiguration.WHITELIST_URLS;
