+++ lsadf_api/src/test/java/com/lsadf/application/bdd/when/BddCurrencyWhenStepDefinitions.java
@@ -21,7 +21,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import com.lsadf.application.bdd.BddLoader;
 import com.lsadf.application.bdd.BddUtils;
-import com.lsadf.application.controllers.CurrencyController;
+import com.lsadf.application.game.currency.CurrencyController;
 import com.lsadf.core.domain.game.currency.Currency;
 import com.lsadf.core.infra.web.config.auth.JwtAuthentication;
 import com.lsadf.core.infra.web.controllers.ControllerConstants;
