+++ lsadf_api/src/test/java/com/lsadf/application/bdd/when/BddStageWhenStepDefinitions.java
@@ -21,7 +21,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import com.lsadf.application.bdd.BddLoader;
 import com.lsadf.application.bdd.BddUtils;
-import com.lsadf.application.controllers.StageController;
+import com.lsadf.application.game.stage.StageController;
 import com.lsadf.core.domain.game.stage.Stage;
 import com.lsadf.core.infra.web.config.auth.JwtAuthentication;
 import com.lsadf.core.infra.web.controllers.ControllerConstants;
