+++ lsadf_api/src/test/java/com/lsadf/application/unit/controllers/CurrencyControllerTests.java
@@ -21,8 +21,8 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.lsadf.application.controllers.CurrencyController;
-import com.lsadf.application.controllers.impl.CurrencyControllerImpl;
+import com.lsadf.application.game.currency.CurrencyController;
+import com.lsadf.application.game.currency.CurrencyControllerImpl;
 import com.lsadf.core.infra.web.controllers.advices.GlobalExceptionHandler;
 import com.lsadf.core.infra.web.requests.game.currency.CurrencyRequest;
 import com.lsadf.core.unit.config.UnitTestConfiguration;
