+++ lsadf_api/src/test/java/com/lsadf/application/unit/controllers/StageControllerTests.java
@@ -21,8 +21,8 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.lsadf.application.controllers.StageController;
-import com.lsadf.application.controllers.impl.StageControllerImpl;
+import com.lsadf.application.game.stage.StageController;
+import com.lsadf.application.game.stage.StageControllerImpl;
 import com.lsadf.core.infra.web.controllers.advices.GlobalExceptionHandler;
 import com.lsadf.core.infra.web.requests.game.stage.StageRequest;
 import com.lsadf.core.unit.config.UnitTestConfiguration;
